PumpSettingsEditor

+++ FreeAPS/Sources/Models/Autosens.swift
@@ -1,5 +1,5 @@
 import Foundation
 
 struct Autosens: JSON {
-    let ratio: Double
+    let ratio: Decimal
 }
